Remove cmd folder
