+++ pkg/p2p/kadcast/utils_test.go
@@ -7,6 +7,8 @@ import (
 	"net"
 	"testing"
 	"time"
+
+	"github.com/dusk-network/dusk-blockchain/pkg/util/container/ring"
 )
 
 func testConv(t *testing.T) {
@@ -42,6 +44,42 @@ func testUDPConn(t *testing.T) {
 		log.Printf("%v", er)
 	}
 	//}()
+	for {
+
+	}
+}
+
+func TestProtocol(t *testing.T) {
+	// Our node info.
+	var port uint16 = 25519
+	ip := [4]byte{62, 57, 180, 247}
+	router := MakeRouter(ip, port)
+
+	// Create buffer.
+	queue := ring.NewBuffer(500)
+
+	// Launch PacketProcessor rutine.
+	go ProcessPacket(queue, &router)
+
+	// Launch a listener for our node.
+	go StartUDPListener("udp", queue, router.myPeerInfo)
+
+	/*// Create BootstrapNodes Peer structs
+	var port1 uint16 = 25519
+	ip1 := [4]byte{157, 230, 219, 77}
+	boot1 := makePeer(ip1, port1)
+	var bootstrapNodes []Peer
+	bootstrapNodes = append(bootstrapNodes[:], boot1)
+
+	// Start Bootstrapping process.
+	err := initBootstrap(&router, bootstrapNodes)
+	if err != nil {
+		log.Fatal("Error during the Bootstrap Process. Job terminated.")
+	}
+
+	// Once the bootstrap succeeded, start the network discovery.
+	startNetworkDiscovery(&router)
+	*/
 	for {
 
 	}
